coreos: rename to flatcar
