+++ installers/flatcar/files/ignition/config.go
