+++ installers/flatcar/files/ignition/storage.go
