+++ installers/flatcar/installer.go
@@ -1,10 +1,10 @@
-package coreos
+package flatcar
 
 import (
 	"strings"
 
 	"github.com/tinkerbell/boots/conf"
-	"github.com/tinkerbell/boots/installers/coreos/files/ignition"
+	"github.com/tinkerbell/boots/installers/flatcar/files/ignition"
 	"github.com/tinkerbell/boots/job"
 )
