+++ installers/flatcar/main_test.go
@@ -1,4 +1,4 @@
-package coreos
+package flatcar
 
 import (
 	"os"
