+++ installers/flatcar/networkd.go
@@ -1,11 +1,11 @@
-package coreos
+package flatcar
 
 import (
 	"net"
 
 	"github.com/tinkerbell/boots/client"
 	"github.com/tinkerbell/boots/conf"
-	"github.com/tinkerbell/boots/installers/coreos/files/unit"
+	"github.com/tinkerbell/boots/installers/flatcar/files/unit"
 	"github.com/tinkerbell/boots/job"
 )
